Extensions of ruby's standard libraries are separated to other files.
